Added my tests to group class ByteBastersGroupTest; removed my personal SwagLabsShopTest class
